backend/requirements.txt
fastapi
fastapi-cache2[redis]
redis
requests
requests-cache
ujson
uvicorn